#338 - Rewrite of caching expiration
Now we use a logarithmic function for cache invalidation, so more the zoom is high,
less the tile is cached a long time.
Then we use a version based on the timestamp of last updated feature of a layer, so
everytime a feature is updated, the layer cache is invalidated.